avoiding overwriting tables catalog

+++ src/pages/mod.rs
@@ -6,7 +6,7 @@ use traits::Serialize;
 pub const PAGE_SIZE: usize = 4096; // 4 KBs
 pub const INVALID_PAGE: PageId = -1;
 
-pub type PageId = isize;
+pub type PageId = i64;
 
 /// A generic page with an underlying array of [`PAGE_SIZE`] bytes
 /// Other pages must implement From<Page> and Into<Page> traits
